--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -18,5 +18,5 @@ jobs:
       - uses: actions/upload-artifact@v4
         with:
           name: spice2x
-          path: src/spice2x/bin/spice2x
+          path: src/spice2x/bin
           if-no-files-found: error