--- a/src/spice2x/Dockerfile
+++ b/src/spice2x/Dockerfile
@@ -1,8 +1,7 @@
 FROM spicetools/deps
 WORKDIR /src
-RUN chown user:user /src
-USER user
-COPY --chown=user:user --from=gitroot . /src/.git
-COPY --chown=user:user . /src/src/spice2x
+
+COPY --from=gitroot . /src/.git
+COPY . /src/src/spice2x
 WORKDIR /src/src/spice2x
 CMD ["./build_all.sh"]